refacto extracting equilibrium transaction card
